introduce START TRANSCATION FOR command
This commit changes the parser to support a new way to start a
transaction---the START TRANSACTION FOR command. The syntax is

```
START TRANSACTION [transaction_type] FOR ([transaction_args])
```

This command is identical to START TRANSACTION (or BEGIN), except it
first registers a parameterized transaction with the underlying
storage engines. The storage engines can use this information to make
scheduling decisions.

The purpose of the transaction type and the transaction args is to
enable us to map a transaction to a scheduling cluster. The type is an
unsigned integer, and the arguments are a list of literals.